% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remarks.R
\name{remark}
\alias{remark}
\alias{remark.}
\title{Create or Add a Remark}
\usage{
remark(key, level, remark, origin = NA_character_)

remark.(., key, level, remark, origin = NA_character_)
}
\arguments{
\item{key}{character: The remark key.}

\item{level}{character: The severity level of your remark.}

\item{remark}{character: The remark itself. This can be a string interpolation
formula.}

\item{origin}{character: The origin of your remark.}

\item{.}{A value which is available for string interpolation, and is returned.}
}
\value{
A remark \code{data.frame}.
}
\description{
There are two versions of this function. The "dot" version is pipe-friendly
and will take as a first argument a value which is then also returned.
}

