% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remarks.R
\name{remarks}
\alias{remarks}
\title{Retrieve Remarks Associated With Key(s)}
\usage{
remarks(..., filter = NULL)
}
\arguments{
\item{...}{character: The keys for which to ger remarks.}

\item{filter}{An optional (one-sided) formula for filtering remarks further,
  see \code{\link[dplyr]{filter_}}.}
}
\value{
A \code{data.frame} with columns \code{Key, Timestamp, Level, Remark,
Origin}.
}
\description{
This will retrieve remarks associated with one or more keys. If no remarks
exist an empty data.frame is returned.
}
\details{
Each row in a remarks \code{data.frame} represent a "remark".
A remark has its "Key", a "Timestamp" (\code{POSIXct}), A severity "Level",
for categorizing types of messages e.g. differentiating
their importance, the "Remark" itself (\code{character}), and
an "Origin" (\code{character} identifying where the remark originated,
e.g. a function or application).
}

